Fixed return type of tes3magicEffectInstance.effectiveMagnitude

+++ MWSE/TES3MagicEffectInstance.cpp
@@ -1,7 +1,7 @@
 #include "TES3MagicEffectInstance.h"
 
 namespace TES3 {
-	int MagicEffectInstance::getEffectiveMagnitude() const {
+	float MagicEffectInstance::getEffectiveMagnitude() const {
 		return magnitude * (1.0f - resistedPercent / 100.0f);
 	}
 }
